pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "seo-automation"
version = "1.0.0"
description = "Enterprise-grade Webflow meta tag automation"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]==0.111.0",
    "uvicorn[standard]==0.29.0",
    "pydantic>=2.7.1",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.4",
    "aiofiles>=23.2.1",
    "python-jose[cryptography]>=3.3.0",
    "slowapi>=0.1.9",
    "structlog>=24.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.sdist]
include = [
    "app/",
    "README.md",
    "Makefile",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "B", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
disallow_untyped_defs = true